@@ -3,30 +3,27 @@
     "version": "1.2.0",
     "private": true,
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.30",
-        "@fortawesome/free-brands-svg-icons": "^5.15.1",
-        "@fortawesome/free-regular-svg-icons": "^5.15.1",
-        "@fortawesome/free-solid-svg-icons": "^5.14.0",
-        "@fortawesome/react-fontawesome": "^0.1.11",
-        "@testing-library/jest-dom": "^4.2.4",
-        "@testing-library/react": "^9.5.0",
-        "@testing-library/user-event": "^7.2.1",
+        "@fortawesome/fontawesome-svg-core": "^6.2.1",
+        "@fortawesome/free-brands-svg-icons": "^6.2.1",
+        "@fortawesome/free-regular-svg-icons": "^6.2.1",
+        "@fortawesome/free-solid-svg-icons": "^6.2.1",
+        "@fortawesome/react-fontawesome": "^0.2.0",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^14.4.3",
         "ansi-regex": "^6.0.1",
-        "array-move": "^3.0.1",
-        "create-react-app": "^5.0.0",
-        "framer-motion": "4.1.17",
-        "gh-pages": "^3.1.0",
-        "netlify-cli": "^8.4.3",
-        "node-sass": "^7.0.0",
-        "prettier": "^2.5.1",
-        "react": "^17.0.2",
-        "react-dom": "^16.13.1",
-        "react-photo-gallery": "^8.0.0",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "^5.0.0",
-        "react-scroll-percentage": "^4.2.0",
-        "react-sortable-hoc": "^1.11.0",
-        "typescript": "^4.0.2"
+        "array-move": "^4.0.0",
+        "create-react-app": "^5.0.1",
+        "framer-motion": "8.0.2",
+        "netlify-cli": "^12.5.0",
+        "node-sass": "^8.0.0",
+        "prettier": "^2.8.1",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
+        "react-router-dom": "^6.6.1",
+        "react-scripts": "^5.0.1",
+        "react-scroll-percentage": "^4.3.0",
+        "typescript": "^4.9.4"
     },
     "scripts": {
         "start": "react-scripts start",
